refactor: use stdOut for all commands as default
This builds the base for using tab completion. Additionally it removes
a lot of manual cobra settings and allows commands to print their help
to stdOut, which is a GNU standard.

The only exception are the promptui-promts. For these it makes most sense that they are being presented on stdErr.