fix: correct handling of multiline output
This fixes an issue where on some shells and terminals the output of the
commands would be displayed as a single line. Apart from not looking
nice, this was also an issue for commands that got sourced.

This fix tackles output by the konf wrapper, konf-go binary output was
always correct.